add test for `create_object()`
