+++ MolecularNodes/obj.py
@@ -10,14 +10,11 @@ def create_object(name: str, collection: bpy.types.Collection, locations, bonds=
     ----------
     name : str
         The name of the mesh object to be created.
-
     collection : bpy.types.Collection
         The collection to which the mesh object will be added.
-
     locations : array-like
         The list of vertex locations for the mesh, an nx3 np array of locations.
         Each element in the list represents a 3D point (x, y, z) for a vertex.
-
     bonds : list of tuples, optional
         The list of vertex index pairs representing bonds as edges for the mesh.
         Each tuple should contain two vertex indices (e.g., (index1, index2)).
@@ -36,11 +33,13 @@ def create_object(name: str, collection: bpy.types.Collection, locations, bonds=
 
     Example
     -------
+    ```python
     # Create a mesh object named "MyMesh" in the collection "MyCollection"
     # with vertex locations and bond edges.
     locations = [[0.0, 0.0, 0.0], [1.0, 0.0, 0.0], [1.0, 1.0, 0.0]]
     bonds = [(0, 1), (1, 2), (2, 0)]
     my_object = create_object("MyMesh", bpy.data.collections['Collection'], locations, bonds)
+    ```
     """
     # create a new mesh
     mol_mesh = bpy.data.meshes.new(name)
@@ -54,29 +53,28 @@ def add_attribute(object: bpy.types.Object, name: str, data, type="FLOAT", domai
     """
     Add an attribute to the given object's geometry on the given domain.
 
-    Parameters:
+    Parameters
+    ----------
         object : bpy.types.Object
             The object to which the attribute will be added.
-
         name : str
             The name of the attribute.
-
         data : array-like
             The data to be assigned to the attribute. For "FLOAT_VECTOR" attributes, it should be a 1D array
             representing the vector data.
-
         type : str, optional, default: "FLOAT"
             The data type of the attribute. Possible values are "FLOAT", "FLOAT_VECTOR", "INT", or "BOOLEAN".
-
         domain : str, optional, default: "POINT"
             The domain to which the attribute is added. Possible values are "POINT" or other domains supported
             by the object.
 
-    Returns:
+    Returns
+    -------
         Any
             The newly created attribute, which can be further manipulated or used in the 3D environment.
 
-    Notes:
+    Notes
+    -----
         - The function supports adding both scalar and vector attributes.
         - The "FLOAT_VECTOR" attribute requires the input data to be a 1D array, and it will be reshaped internally
           to represent vectors with 3 components (x, y, z).
@@ -102,15 +100,14 @@ def get_attribute(obj: bpy.types.Object, att_name='position') -> np.array:
     ----------
     obj : bpy.types.Object
         The Blender object from which the attribute will be retrieved.
-
     att_name : str, optional
         The name of the attribute to retrieve. Default is 'position'.
-
+        
     Returns
     -------
     np.array
         The attribute data as a NumPy array.
-
+    
     Notes
     -----
     - This function retrieves the specified attribute from the object and returns it as a NumPy array.
@@ -118,8 +115,10 @@ def get_attribute(obj: bpy.types.Object, att_name='position') -> np.array:
 
     Example
     -------
+    ```python
     # Assuming 'my_object' is a Blender object with an attribute named 'my_attribute'
     attribute_data = get_attribute(my_object, 'my_attribute')
+    ```
     """
 
     # Get the attribute from the object's mesh
